# mul and from optimizations