- mul is ~30% faster
- from_little_endian / from_big_endian is ~80% faster 

before:

```
test u128_mul      ... bench:     151,032 ns/iter (+/- 32,331)
test u256_from_be  ... bench:          57 ns/iter (+/- 0)
test u256_from_le  ... bench:           51 ns/iter (+/- 1)
test u256_full_mul ... bench:     228,725 ns/iter (+/- 30,183)
test u256_mul      ... bench:     123,390 ns/iter (+/- 32,967)
```

after: 

```
test u128_mul      ... bench:      98,213 ns/iter (+/- 13,330)
test u256_from_be  ... bench:          13 ns/iter (+/- 0)
test u256_from_le  ... bench:           7 ns/iter (+/- 0)
test u256_full_mul ... bench:     199,849 ns/iter (+/- 19,195)
test u256_mul      ... bench:     105,883 ns/iter (+/- 14,433)
```

pleasereview @NikVolf / @Vurich